requirements.txt
streamlit
python-dotenv
requests
stability-sdk